Support int64 literal goldens

+++ tests/golden/outputs/wgsl/conversion-float-to-int-no-f64.wgsl
@@ -14,12 +14,12 @@ fn test_f16_to_u32(f: f16) -> u32 {
     return u32(f);
 }
 
-fn test_f16_to_i64(f: f16) -> i32 {
-    return i32(f);
+fn test_f16_to_i64(f: f16) -> i64 {
+    return i64(f);
 }
 
-fn test_f16_to_u64(f: f16) -> u32 {
-    return u32(f);
+fn test_f16_to_u64(f: f16) -> u64 {
+    return u64(f);
 }
 
 fn test_f32_to_i32(f: f32) -> i32 {
@@ -30,12 +30,12 @@ fn test_f32_to_u32(f: f32) -> u32 {
     return u32(f);
 }
 
-fn test_f32_to_i64(f: f32) -> i32 {
-    return i32(f);
+fn test_f32_to_i64(f: f32) -> i64 {
+    return i64(f);
 }
 
-fn test_f32_to_u64(f: f32) -> u32 {
-    return u32(f);
+fn test_f32_to_u64(f: f32) -> u64 {
+    return u64(f);
 }
 
 fn test_f16_to_i32_vec(f: vec2<f16>) -> vec2<i32> {
@@ -46,12 +46,12 @@ fn test_f16_to_u32_vec(f: vec2<f16>) -> vec2<u32> {
     return vec2<u32>(f);
 }
 
-fn test_f16_to_i64_vec(f: vec2<f16>) -> vec2<i32> {
-    return vec2<i32>(f);
+fn test_f16_to_i64_vec(f: vec2<f16>) -> vec2<i64> {
+    return vec2<i64>(f);
 }
 
-fn test_f16_to_u64_vec(f: vec2<f16>) -> vec2<u32> {
-    return vec2<u32>(f);
+fn test_f16_to_u64_vec(f: vec2<f16>) -> vec2<u64> {
+    return vec2<u64>(f);
 }
 
 fn test_f32_to_i32_vec(f: vec2<f32>) -> vec2<i32> {
@@ -62,12 +62,12 @@ fn test_f32_to_u32_vec(f: vec2<f32>) -> vec2<u32> {
     return vec2<u32>(f);
 }
 
-fn test_f32_to_i64_vec(f: vec2<f32>) -> vec2<i32> {
-    return vec2<i32>(f);
+fn test_f32_to_i64_vec(f: vec2<f32>) -> vec2<i64> {
+    return vec2<i64>(f);
 }
 
-fn test_f32_to_u64_vec(f: vec2<f32>) -> vec2<u32> {
-    return vec2<u32>(f);
+fn test_f32_to_u64_vec(f: vec2<f32>) -> vec2<u64> {
+    return vec2<u64>(f);
 }
 
 fn test_const_eval() {
@@ -75,26 +75,26 @@ fn test_const_eval() {
     var max_f16_to_i32: i32 = i32(MAX_F16);
     var min_f16_to_u32: u32 = u32(MIN_F16);
     var max_f16_to_u32: u32 = u32(MAX_F16);
-    var min_f16_to_i64: i32 = i32(MIN_F16);
-    var max_f16_to_i64: i32 = i32(MAX_F16);
-    var min_f16_to_u64: u32 = u32(MIN_F16);
-    var max_f16_to_u64: u32 = u32(MAX_F16);
+    var min_f16_to_i64: i64 = i64(MIN_F16);
+    var max_f16_to_i64: i64 = i64(MAX_F16);
+    var min_f16_to_u64: u64 = u64(MIN_F16);
+    var max_f16_to_u64: u64 = u64(MAX_F16);
     var min_f32_to_i32: i32 = i32(MIN_F32);
     var max_f32_to_i32: i32 = i32(MAX_F32);
     var min_f32_to_u32: u32 = u32(MIN_F32);
     var max_f32_to_u32: u32 = u32(MAX_F32);
-    var min_f32_to_i64: i32 = i32(MIN_F32);
-    var max_f32_to_i64: i32 = i32(MAX_F32);
-    var min_f32_to_u64: u32 = u32(MIN_F32);
-    var max_f32_to_u64: u32 = u32(MAX_F32);
+    var min_f32_to_i64: i64 = i64(MIN_F32);
+    var max_f32_to_i64: i64 = i64(MAX_F32);
+    var min_f32_to_u64: u64 = u64(MIN_F32);
+    var max_f32_to_u64: u64 = u64(MAX_F32);
     var min_abstract_float_to_i32: i32 = i32(MIN_ABSTRACT_FLOAT);
     var max_abstract_float_to_i32: i32 = i32(MAX_ABSTRACT_FLOAT);
     var min_abstract_float_to_u32: u32 = u32(MIN_ABSTRACT_FLOAT);
     var max_abstract_float_to_u32: u32 = u32(MAX_ABSTRACT_FLOAT);
-    var min_abstract_float_to_i64: i32 = i32(MIN_ABSTRACT_FLOAT);
-    var max_abstract_float_to_i64: i32 = i32(MAX_ABSTRACT_FLOAT);
-    var min_abstract_float_to_u64: u32 = u32(MIN_ABSTRACT_FLOAT);
-    var max_abstract_float_to_u64: u32 = u32(MAX_ABSTRACT_FLOAT);
+    var min_abstract_float_to_i64: i64 = i64(MIN_ABSTRACT_FLOAT);
+    var max_abstract_float_to_i64: i64 = i64(MAX_ABSTRACT_FLOAT);
+    var min_abstract_float_to_u64: u64 = u64(MIN_ABSTRACT_FLOAT);
+    var max_abstract_float_to_u64: u64 = u64(MAX_ABSTRACT_FLOAT);
 }
 
 @compute
